GILL-7: implementation of fsd architecture
