+++ config/mainBuild/buildResolvers.ts
@@ -1,7 +1,16 @@
 import { ResolveOptions } from "webpack";
+import { BuildOptions } from './types/config';
 
-export function buildResolvers(): ResolveOptions {
+export function buildResolvers(options: BuildOptions): ResolveOptions {
     return {
         extensions: ['.tsx', '.ts', '.js'],
-    }
+    };
 }
+
+// preferAbsolute: true,
+// modules: [options.paths.src, 'node_modules'],
+// mainFields: ['index'],
+// alias: {},
+// fallback: {
+//     "events": false,
+// },
