+++ config/mainBuild/types/config.ts
@@ -1,19 +1,20 @@
 export type BuildMode = 'production' | 'development';
 
-export type BuildPaths =  {
+export type BuildPaths = {
     entry: string,
     build: string,
     html: string,
-}
+    src: string,
+};
 
 export type BuildOptions = {
     mode: BuildMode,
     paths: BuildPaths,
     isDev: boolean,
     port: number,
-}
+};
 
 export type BuildEnv = {
     mode: BuildMode,
     port: number,
-}
+};
