+++ src/app/providers/themeProvider/index.ts
@@ -1,0 +1,7 @@
+import { ThemeProvider } from './ui/ThemeProvider';
+import { useTheme } from './lib/useTheme';
+
+export {
+    ThemeProvider,
+    useTheme,
+};
