+++ src/app/styles/variables/globals.scss
