+++ src/app/types/global.d.ts
@@ -1,0 +1,7 @@
+declare module '*.scss' {
+    interface IClassNames {
+        [className: string]: string;
+    }
+    const classNames: IClassNames;
+    export = classNames;
+}
