--- src/components/counter/Counter.module.scss
@@ -1,7 +1,0 @@
-.wrapper {
-    display: block;
-}
-
-.buttonCounter {
-    color: green;
-}
