--- src/global.d.ts
@@ -1,7 +1,0 @@
-declare module '*.scss' {
-  interface IClassNames {
-    [className: string]: string;
-  }
-  const classNames: IClassNames;
-  export = classNames;
-}
