+++ src/index.tsx
@@ -1,8 +1,9 @@
 import { render } from 'react-dom';
 import { BrowserRouter } from 'react-router-dom';
 
-import App from './App';
-import { ThemeProvider } from './theme/ThemeProvider';
+import App from './app/App';
+import { ThemeProvider } from './app/providers/themeProvider';
+
 
 render(
     <BrowserRouter>
