+++ webpack.config.ts
@@ -6,23 +6,24 @@ import { BuildEnv, BuildPaths } from './config/mainBuild/types/config';
 
 
 export default (env: BuildEnv) => {
-    const {mode = 'development', port = 3000} = env;
+    const { mode = 'development', port = 3000 } = env;
 
     const paths: BuildPaths = {
         entry: path.resolve(__dirname, 'src', 'index.tsx'),
         build: path.resolve(__dirname, 'build'),
         html: path.resolve(__dirname, 'public', 'index.html'),
-    }
-    
+        src: path.resolve(__dirname, 'src'),
+    };
+
     const isDev = mode === 'development';
     const PORT = port;
-    
+
     const config: webpack.Configuration = buildWebpackConfig({
         mode: mode,
         paths: paths,
         isDev: isDev,
         port: PORT,
-    })
+    });
 
     return config;
 };
